requirements.txt
discord
python-dotenv
langchain
langchain_google_genai
tavily-python